Reorganize biotek/ into plate_readers/ subpackages, fix review issues
- Move BioTekBackend -> plate_readers/base.py (shared base)
- Move Cytation code -> plate_readers/cytation/ (base, cytation1, cytation5,
  microscopy_backend, aravis_camera)
- Move SynergyH1 -> plate_readers/synergy/
- Update all internal imports, legacy adapters, and docs (rst + notebooks)

Review fixes:
- Wrap led_off() in try/except inside capture() finally so stop_acquisition
  always runs even if the LED command fails
- Restore asyncio.sleep(0.1) stage settle delay in set_position
- Restore max_image_read_attempts default to 50
- Accept backend_params kwarg in CytationMicroscopyBackend._on_setup
- Add PyGObject to cytation-microscopy extra (needed for Aravis)
- Add autofocus example to cytation hello-world notebook
