+++ pylabrobot/agilent/biotek/__init__.py
@@ -1,7 +1,11 @@
-from .biotek import BioTekBackend
-from .cytation1 import Cytation1
-from .cytation5 import Cytation5
-from .cytation_microscopy_backend import CytationImagingConfig, CytationMicroscopyBackend
+from .plate_readers import (
+  BioTekBackend,
+  Cytation1,
+  Cytation5,
+  CytationImagingConfig,
+  CytationMicroscopyBackend,
+  SynergyH1,
+  SynergyH1Backend,
+)
 from .loading_tray_backend import BioTekLoadingTrayBackend
 from .el406 import EL406, EL406Driver, EL406PlateWasher96Backend, EL406ShakingBackend
-from .synergy_h1 import SynergyH1, SynergyH1Backend
