+++ pylabrobot/agilent/biotek/plate_readers/cytation/aravis_camera.py
