+++ pylabrobot/agilent/biotek/plate_readers/synergy/__init__.py
@@ -1,0 +1,1 @@
+from .synergy_h1 import SynergyH1, SynergyH1Backend
